Move metrics tracker into shared utilities
