+++ src/shared/metrics-tracker.js
@@ -1,0 +1,1 @@
+export { createMetricsTracker } from "./utils/metrics-tracker.js";
